added: qr code for url scanning in table and homepage
