+++ next.config.js
@@ -1,0 +1,6 @@
+const nextConfig = {
+    // In strict react will render two times. To catch error early.
+    // reactStrictMode: false
+};
+
+module.exports = nextConfig;
