+++ package.json
@@ -13,6 +13,7 @@
         "@radix-ui/react-alert-dialog": "1.0.4",
         "@radix-ui/react-avatar": "^1.0.4",
         "@radix-ui/react-checkbox": "^1.0.4",
+        "@radix-ui/react-dialog": "^1.0.5",
         "@radix-ui/react-dropdown-menu": "^2.0.6",
         "@radix-ui/react-label": "^2.0.2",
         "@radix-ui/react-menubar": "^1.0.4",
@@ -23,6 +24,7 @@
         "bcrypt": "^5.1.1",
         "class-variance-authority": "^0.7.0",
         "clsx": "^2.0.0",
+        "lodash": "^4.17.21",
         "lucide-react": "^0.279.0",
         "next": "^13.5.3",
         "next-auth": "4.23.1",
@@ -30,6 +32,8 @@
         "react": "^18.2.0",
         "react-dom": "^18.2.0",
         "react-hook-form": "^7.47.0",
+        "react-hot-toast": "^2.4.1",
+        "react-qr-code": "^2.0.12",
         "tailwind-merge": "^1.14.0",
         "tailwindcss-animate": "^1.0.7",
         "type-fest": "^4.6.0",
@@ -39,6 +43,7 @@
         "@trivago/prettier-plugin-sort-imports": "^4.2.0",
         "@types/bcrypt": "^5.0.0",
         "@types/cookie": "^0.5.2",
+        "@types/lodash": "^4.14.202",
         "@types/node": "20.7.1",
         "@types/react": "18.2.23",
         "@typescript-eslint/eslint-plugin": "^6.7.3",
