+++ src/app/(mainpage)/page.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { HomePage } from '@/components/homepage';
 
 const MainPage = () => {
