+++ src/app/layout.tsx
@@ -1,4 +1,7 @@
 import '../styles/globals.css';
+import { Session } from 'next-auth';
+import getServerSession from '@/lib/getServerSession';
+import ClientSessionProvider from '@/common/provider/client-session';
 
 export const metadata = {
     title: 'Shorty URL',
@@ -6,9 +9,15 @@ export const metadata = {
 };
 
 const RootLayout = async ({ children }: { children: React.ReactNode }) => {
+    const session: Session = await getServerSession();
+
     return (
         <html lang="en">
-            <body className="w-screen h-screen">{children}</body>
+            <body className="w-screen h-screen">
+                <ClientSessionProvider session={session}>
+                    {children}
+                </ClientSessionProvider>
+            </body>
         </html>
     );
 };
