+++ src/common/enums/style-enums.ts
@@ -1,0 +1,10 @@
+export const ComponentSize = {
+    LARGE: 'large',
+    MEDIUM: 'medium',
+    SMALL: 'small',
+} as const;
+
+export const ComponentIntent = {
+    PRIMARY: 'primary',
+    SECONDARY: 'SECONDARY',
+} as const;
