+++ src/components/ui/index.ts
@@ -1,11 +1,9 @@
-export * from './form'
-export * from './input'
-export * from './button'
-export * from './label'
-export * from './alert-dialog'
-export * from './menubar'
-export * from './navigation-menu'
-export * from './table'
-export * from './avatar'
-export * from './use-toast'
-export * from './toaster'
+export * from './form';
+export * from './input';
+export * from './button';
+export * from './label';
+export * from './alert-dialog';
+export * from './menubar';
+export * from './navigation-menu';
+export * from './table';
+export * from './avatar';
