--- src/components/url-data-table/index.ts
@@ -1,1 +1,0 @@
-export * from './url-data-table';
