--- src/components/url-data-table/types.ts
@@ -1,13 +1,0 @@
-export type ShortendUrlResponseData = () => {
-    id: string;
-    urlId: string;
-    shortUrl: string;
-    longUrl: string;
-    createdAt: string;
-};
-
-export type UrlData = {
-    urlId: string;
-    shortUrl: string;
-    longUrl: string;
-};
