+++ src/components/url-history-table/index.ts
@@ -1,0 +1,1 @@
+export * from './url-data-table';
