+++ src/components/url-history-table/types.ts
@@ -1,0 +1,13 @@
+export type ShortendUrlResponseData = () => {
+    id: string;
+    urlId: string;
+    shortUrl: string;
+    longUrl: string;
+    createdAt: string;
+};
+
+export type UrlData = {
+    urlId: string;
+    shortUrl: string;
+    longUrl: string;
+};
