+++ src/lib/getServerSession.ts
@@ -1,8 +1,7 @@
-import { getServerSession as _getServerSession } from 'next-auth/next'
-import { cookies, headers } from 'next/headers'
-
-import { authOptionsWrapper } from '@/app/api/auth/[...nextauth]/route'
-import { Session } from 'next-auth'
+import { getServerSession as _getServerSession } from 'next-auth/next';
+import { cookies, headers } from 'next/headers';
+import { authOptionsWrapper } from '@/app/api/auth/[...nextauth]/route';
+import { Session } from 'next-auth';
 
 const getServerSession = (): Promise<Session> => {
     return _getServerSession(
@@ -13,7 +12,7 @@ const getServerSession = (): Promise<Session> => {
             } as any,
             { params: { nextauth: ['session'] } }
         )[2]
-    )
-}
+    );
+};
 
-export default getServerSession
+export default getServerSession;
